// client/src/components/layouts/layout.types.tsx
export type JustifyContent =
    | "flex-start"
    | "flex-end"
    | "space-between"
    | "space-evenly"
    | "space-around"
    | "inherit"
    | "center"
    | "initial";

export type AlignItems =
    | "flex-start"
    | "flex-end"
    | "center"
    | "baseline"
    | "stretch"
    | "normal"
    | "inherit"
    | "initial";

export type FlexDirection = "row" | "column";

export type Overflow =
    | "scroll"
    | "none"
    | "clip"
    | "visible"
    | "hidden"
    | "auto";